Refactor: make config simpler, refactor if else
